Moved types for Parse to package "type". Some refactor in Parse
